miniapps/comparator/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0 